+++ NEWS.md
@@ -1,3 +1,6 @@
+# MachineLearning 0.0.3.9000 DEVEL VERSION
+* 
+
 # MachineLearning 0.0.3
 * Fix a bug in CART function