+++ NEWS.md
@@ -1,5 +1,8 @@
-# MachineLearning 0.0.3.9000 DEVEL VERSION
-* 
+# MachineLearning 0.0.3.9002 DEVEL VERSION
+* Refix CART function plot & summary method
+* Fix consistence of MLA S3methods 
+* New indepedent function for CART plot with compatibility with Rpart plots (Unstable version)
+
 
 # MachineLearning 0.0.3
 * Fix a bug in CART function